backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1

# AI/ML Libraries
tensorflow==2.15.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# NLP and Transformers
transformers==4.37.2
torch==2.2.0
sentence-transformers==2.3.1
langchain==0.1.7
langchain-openai==0.0.6

# Translation
deep-translator==1.11.4
indic-transliteration==2.3.37

# Data Processing
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Utilities
requests==2.31.0
aiofiles==23.2.1
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Documentation
mkdocs==1.5.3
email-validator>=2.0.0